requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.5
django-allauth==0.36.0
django-rest-auth==0.9.3
django-sslserver==0.20
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.6
oauthlib==2.1.0
Pillow==5.1.0
psycopg2==2.7.4
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
django-cors-headers==2.2.0
xmtodict==0.11.0